requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
openai==2.17.0
httpx==0.27.2
redis==5.2.1
python-docx==1.1.2
pydantic==2.10.3
pydantic-settings==2.6.1
aiofiles==24.1.0
python-multipart==0.0.20
tiktoken==0.8.0
cachetools==5.5.0
python-dateutil==2.9.0
openpyxl==3.1.5
python-dotenv==1.0.1
pyodbc>=5.0.1
boto3>=1.34.0
reportlab>=4.0
sse-starlette>=1.6.0
structlog>=24.1.0
slowapi>=0.1.9
